Moved multichain tests to Tribler/Test/Community/Multichain directory
