+++ Tribler/Test/Community/Multichain/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for the Multichain community in Tribler.
+"""
